Refactor game logic, fixed player not switching on the last move
The game logic for the Tic Tac Toe project has been significantly refactored. Changes include an update to the structure and datatype of the game board, the elimination of redundant logic in determining valid moves, and the addition of clearer and more efficient methods for checking which player has won or if the game is a draw. Player is switched right after a valid move is performed to avoid bugs. The UI has been adjusted to reflect these changes, simplifying the button click method and adding a grid calculation for the chosen move.