Fixed CheckForWin method, Refactored GameLogic
Updated the GameLogic and GameBoard classes to enhance the efficiency and readability of the code. This included changing variable names to comply with naming conventions, adding Console output for debugging CheckForWin method, it was always return false even when there is a winner, the issue was with the casting of the enum to char. casting an enum to a char would not give you the char representation ('X', 'O') as one would expect. It would instead give you a char equivalent of the numeric value backing the specific enum label. If unspecified, the numeric values are assigned incrementally starting from 0, so PlayerSymbol entries were updated to include their ASCII equivalent for easier comparison.